requirements.txt
# TownSquare - Civic Issue Reporting System
# AI-powered complaint analysis using Google Gemini Pro

# Core AI Service
google-generativeai>=0.3.0

# Django dependencies
Django==5.2.4
djangorestframework==3.15.0

# Image processing
Pillow>=10.4.0

# HTTP requests and utilities
requests>=2.31.0
python-decouple==3.8

# UI components
django-crispy-forms==2.1
crispy-bootstrap5==0.7

# Optional: Hugging Face as backup (uncomment if needed)
# transformers==4.35.0
# torch==2.1.0


